requirements.txt
Django<2.0
Pillow
requests
